Many namespace renames, file renames and organisation to tidy up the code

+++ Pocketailor/Model/Helpers.cs
@@ -55,11 +55,11 @@ namespace Pocketailor.Model
         }
 
 
-        public static double GetUnixTime()
+        public static long GetUnixTime()
         {
             DateTime t0 = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
             double deltaSeconds = (DateTime.UtcNow - t0).TotalSeconds;
-            return deltaSeconds;
+            return (long)deltaSeconds;
         }
